Move plain text to Verses container from Text container
