+++ app/containers/HomePage/index.js
@@ -561,7 +561,7 @@ class HomePage extends React.PureComponent {
 		const { userId, userAuthenticated } = this.props.profile;
 
 		const autoPlayEnabled = userSettings.get('autoPlayEnabled');
-		const { isScrollingDown, footerDistance: distance } = this.state;
+		const { isScrollingDown } = this.state;
 		const { userNotes, bookmarks, text: updatedText } = this.props.textData;
 		const token = this.props.homepage.match.params.token || '';
 		const verse = this.props.homepage.match.params.verse || '';
@@ -603,43 +603,24 @@ class HomePage extends React.PureComponent {
 					)}
 					<Text
 						books={books}
-						userId={userId}
 						text={updatedText}
 						hasVideo={hasVideo}
-						distance={distance}
 						verseNumber={verse}
 						userNotes={userNotes}
 						bookmarks={bookmarks}
-						bibleId={activeTextId}
 						menuIsOpen={isMenuOpen}
 						highlights={highlights}
-						audioSource={audioSource}
 						activeTextId={activeTextId}
 						activeBookId={activeBookId}
 						loadingAudio={loadingAudio}
 						userSettings={userSettings}
 						activeChapter={activeChapter}
-						textDirection={textDirection}
-						invalidBibleId={invalidBibleId}
-						activeBookName={activeBookName}
-						notesActive={isNotesModalActive}
 						formattedSource={formattedSource}
 						videoPlayerOpen={videoPlayerOpen}
 						isScrollingDown={isScrollingDown}
 						audioPlayerState={audioPlayerState}
-						userAuthenticated={userAuthenticated}
-						informationActive={isInformationModalActive}
 						loadingNewChapterText={loadingNewChapterText}
 						chapterTextLoadingState={chapterTextLoadingState}
-						addBookmark={this.addBookmark}
-						addHighlight={this.addHighlight}
-						setActiveNote={this.setActiveNote}
-						getCopyrights={this.getCopyrights}
-						toggleNotesModal={this.toggleNotesModal}
-						deleteHighlights={this.deleteHighlights}
-						setActiveNotesView={this.setActiveNotesView}
-						setTextLoadingState={this.setTextLoadingState}
-						toggleInformationModal={this.toggleInformationModal}
 					/>
 				</div>
 				<TransitionGroup>
