+++ app/utils/highlightingUtils.js
@@ -181,6 +181,31 @@ const getTextInSelectedNodes = ({
 	);
 };
 
+// Because the system captures the verse numbers this needs to be used
+// Can accept plainText boolean as third arg if needed
+const calcDistance = (last, first /* plainText */) => {
+	// If the last verse is equal to the first verse then I don't need a diff
+	if (last === first) return 0;
+	let stringDiff = '';
+
+	for (let i = first + 1; i <= last; i += 1) {
+		// Adds the length of each verse number
+		stringDiff += i.toFixed(0);
+		// Adds characters for text to account for spaces in verse numbers
+		stringDiff += '11';
+	}
+	// Gets the total length of the distance neededk
+
+	return stringDiff.length;
+};
+
+const getReference = (verseStart, verseEnd, activeBookName, activeChapter) =>
+	`${activeBookName} ${activeChapter}:${
+		verseStart === verseEnd || !verseEnd
+			? verseStart
+			: `${verseStart}-${verseEnd}`
+	}`;
+
 export {
 	getFormattedParentVerseNumber,
 	getFormattedParentVerse,
@@ -193,4 +218,6 @@ export {
 	getClosestParent,
 	getOffsetNeededForPsalms,
 	replaceCharsRegex,
+	calcDistance,
+	getReference,
 };
